# doc/api.yaml
openapi: 3.0.2
info:
  title: WASAPhoto
  description: |
    OpenAPI document for the WASAPhoto application
  version: 1.1.0  
paths:   
  /users:
    post:

      tags: ["users"]
      summary: Logs in the user
      description: |
        If the user does not exist, it will be created, and an identifier is returned.
        If the user exist the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User detail (the username)
        content:
          application/json:
            schema:
              description: The username of the current user to login
              type: object
              properties:
                request:
                  $ref: "#/components/schemas/Username"
        required: true      
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "201":
          $ref: "#/components/responses/Created"
        "400": 
          $ref: '#/components/responses/BadRequest'
        "401": 
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalServerError'
      security: [] 
    get:
      tags: ["users"]
      summary: Get the user profile by username
      description: |
        Get the profile of a user by username
      operationId: getUsers

      parameters:
        - name: username
          in: query
          description: The username of the user
          required: true
          schema:
            $ref: "#/components/schemas/Username"

      responses:
        "200":
          description: The user has been found
          content:
            application/json:
              schema: 
                description: List of username found
                type: object
                properties:
                  users:
                    description: The list of users found
                    type: array
                    minItems: 0
                    maxItems: 50
                    items:
                      $ref: "#/components/schemas/Username"
        "400": 
          $ref: '#/components/responses/BadRequest'
        "401": 
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /users/{idUser}:
    parameters:
        - $ref: "#/components/parameters/idUser"
    get:
      tags: ["users"]
      summary: Get the user's profile
      description: Return the profile of a user
      operationId: getUserProfile
      responses:
        "200":
          description: Profile returned
          content:
            application/json:
              schema: 
                description: The profile of the user
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "400": 
          $ref: '#/components/responses/BadRequest'
        "401": 
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: ["users"]
      summary: Change the username
      description: Set a new username
      operationId: setMyUserName
      requestBody:
        description: The new username
        content:
          application/json:
            schema:
              description: The new username
              type: object
              properties:
                username:
                  $ref: "#/components/schemas/Username"
        required: true
      responses:
        "200":
          description: Username updated
          content:
            application/json:
              schema:
                description: The new username
                type: object
                properties:
                  username:
                    $ref: "#/components/schemas/Username"
        "400": 
          $ref: '#/components/responses/BadRequest'
        "401": 
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /users/{idUser}/stream:
    parameters:
      - $ref: "#/components/parameters/idUser"
    get:
      tags: ["users"]
      summary: Get my stream
      description: |
        Get my post stream. It will contains post of my followers, in reverse chronological order
      operationId: getMyStream
      responses:
        "200": 
          description: Stream returned
          content:
            application/json:
              schema:
                description: The stream of the user
                type: object
                properties:
                  posts:
                    description: The list of posts
                    type: array
                    minItems: 0
                    maxItems: 50
                    items:
                      $ref: "#/components/schemas/Post"
        "400": 
          $ref: '#/components/responses/BadRequest'
        "401": 
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /users/{idUser}/followers:
    parameters:
      - $ref: "#/components/parameters/idUser"
    get:
      tags: ["follows"]
      summary: List all the followers
      operationId: listFollowers
      description: |
        This method allow to list all the followers of the specified user
      responses:
        "200":  
          description: Followers returned
          content:
            application/json:
              schema:
                description: The list of followers
                type: object
                properties:
                  followers:
                    description: The list of users who follow the specified user
                    type: array
                    minItems: 0
                    maxItems: 50
                    items:
                      $ref: "#/components/schemas/User"
                    
        "400": 
          $ref: '#/components/responses/BadRequest'
        "401": 
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
  /users/{idUser}/banList:
    parameters:
      - $ref: "#/components/parameters/idUser"
    get:
      tags: ["bans"]
      operationId: getBannedList
      summary: Gets the list of user banned
      description: Gets the list of all users banned by the user idUser
      responses:
        "200":
          description: Banned list returned
          content:
            application/json:
              schema:
                description: The list of banned users
                type: object
                properties:
                  banned:
                    description: The list of banned users
                    type: array
                    minItems: 0
                    maxItems: 50
                    items:
                      $ref: "#/components/schemas/User"
        "400": 
          $ref: '#/components/responses/BadRequest'
        "401": 
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
  /users/{idUser}/posts:
    parameters:
      - $ref: "#/components/parameters/idUser"
    post:
      tags: ["photos"]
      summary: Upload a photo
      description: | 
        Upload a post made up of an image and other informations (likes, comments, caption, ...)
      operationId: uploadPhoto
      parameters:
        - name: caption
          in: query
          description: Optional description
          required: true
          schema:
            description: The caption of the post
            type: object
            properties:
              caption:
                $ref: "#/components/schemas/Message"
      requestBody:
        description: The binary image file of the photo
        content:
          image/*:    # Accept all type of images
            schema:
              description: The image
              type: object
              properties:
                image:
                  $ref: "#/components/schemas/Image"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400": 
          $ref: '#/components/responses/BadRequest'
        "401": 
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: "#/components/responses/InternalServerError" 
    get:
      tags: ["photos"]
      summary: Get photos
      description: Get the photos posted by the user
      operationId: getUserPhotos
      responses:
        "200":
          description: Photos of the user returned
          content:
            application/json:
              schema:
                description: The list of photos of the user
                type: object
                properties:
                  photos:
                    description: The list of photos posted by the user
                    type: array
                    items:
                      $ref: "#/components/schemas/Post"
                    minItems: 0
                    maxItems: 50
        "400":
          $ref: "#/components/responses/BadRequest"
        "401": 
          $ref: '#/components/responses/Unauthorized'
        "404": 
          $ref: "#/components/responses/NotFound" 
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /users/{idUser}/posts/{idPhoto}/comments:
    parameters:
      - $ref: "#/components/parameters/idUser"
      - $ref: "#/components/parameters/idPhoto"
    post:
      tags: ["comments"]
      summary: Post a comment
      description: | 
        Post a comment on another user's post or even himself
      operationId: commentPhoto
      requestBody:
        description: The caption of the comment
        content:
          application/json:
            schema:
              description: The caption of the comment posted
              type: object
              properties:
                caption:
                  $ref: "#/components/schemas/Comment/properties/caption"
      responses:
        "201":
          $ref: "#/components/responses/Created" 
        "400": 
          $ref: "#/components/responses/BadRequest" 
        "401": 
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: "#/components/responses/Forbidden"
        "500": 
           $ref: "#/components/responses/InternalServerError" 
    get:
      tags: ["comments"]
      summary: Get comments
      description: Get the list of comments posted by user
      operationId: getComments
      responses:
        "200":
          description: Comments returned
          content:
            application/json:
              schema:
                description: The list of comments
                type: object
                properties:
                  comments:
                    description: The list of comments posted by the user
                    type: array
                    minItems: 0
                    maxItems: 50
                    items:
                      $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /users/{idUser}/posts/{idPhoto}/likes:
    parameters:
      - $ref: "#/components/parameters/idUser"
      - $ref: "#/components/parameters/idPhoto"
    put:
      summary: Like a photo
      description: The user put a like on a photo
      operationId: likePhoto
      tags: ["likes"]
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "201":
          $ref: "#/components/responses/Created"
        "400": 
          $ref: "#/components/responses/BadRequest" 
        "401": 
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: "#/components/responses/Forbidden"
        "500": 
           $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Unlike a photo
      description: Remove a like from a photo
      tags: ["likes"]
      operationId: unlikePhoto
      responses:
        "204":
          $ref: "#/components/responses/No content"
        "400": 
          $ref: "#/components/responses/BadRequest" 
        "401": 
          $ref: '#/components/responses/Unauthorized'
        "404": 
          $ref: "#/components/responses/NotFound" 
        "500":
          $ref: "#/components/responses/InternalServerError" 
    get:
      summary: Get likes
      description: Get the list of likes put by the user
      tags: ["likes"]
      operationId: getLikes
      responses:
        "200":
          description: Likes returned
          content:
            application/json:
              schema:
                description: The list of likes
                type: object
                properties:
                  likes:
                    description: The list of likes put by the user
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                    minItems: 0
                    maxItems: 50
        "400":
          $ref: "#/components/responses/BadRequest"
        "401": 
          $ref: '#/components/responses/Unauthorized'
        "404": 
          $ref: "#/components/responses/NotFound" 
        "500":
          $ref: "#/components/responses/InternalServerError" 
components:
  schemas:  
    Image:
      title: Image
      description: This object represents a single image (as binary of maximum size 20MB)
      type: string
      format: binary
      minLength: 1
      maxLength: 20971520
    Post:
      title: Post
      description: This object represent a single post, with all the related informations
      type: object
      properties:
        id:
          description: The ID of the post
          type: integer
          readOnly: true
          example: 100
        image:
          $ref: "#/components/schemas/Image"
        user-id:
          $ref: "#/components/schemas/User/properties/id"
        username:
          $ref: "#/components/schemas/Username"
        like-numbers:
          description: Number of likes of this post
          type: integer
          minimum: 0
          example: 100
        comment-numbers:
          description: Number of commments of this post
          type: integer
          minimum: 0
          example: 20 
        upload-time:
          $ref: "#/components/schemas/DateTime"
        liked:
          description: If the post is liked
          type: boolean
          example: true
    User:
      title: User
      description: |
        This object reperesents all the info related to a user
      type: object
      properties:
        id:
          description: The ID of the user
          type: integer
          readOnly: true
          example: 100
        name:
          description: User's name
          type: string
          pattern: '^.*?$'
          minLength: 2
          maxLength: 52
          example: Andrea
        surname:
          description: User's surname
          type: string
          pattern: '^.*?$'
          minLength: 2
          maxLength: 52
          example: Rossi
        username:
          $ref: "#/components/schemas/Username"
        postNo:
          description: The number of post
          type: integer
          minimum: 0
          example: 200
        followerNo:
          description: The number of follower
          type: integer
          minimum: 0
          example: 25
        followingNo:
          description: The number of followed
          type: integer
          minimum: 0
          example: 10
        followed:
          description: If the user is followed
          type: boolean
          example: false
    Username:
      title: Username
      description: This object represent the user name
      type: object
      properties:
        name:
          description: The username
          type: string
          pattern: '^[a-zA-Z0-9_\.#$]+$'
          minLength: 3
          maxLength: 16
          example: andremuso
    Comment:
      title: Comment
      description: This object represent a comment on a post
      type: object
      properties:
        id:
          description: The ID of the comment
          type: integer
          readOnly: true
          example: 100
        userId:
          $ref: "#/components/schemas/User/properties/id"
        photoId:
          $ref: "#/components/schemas/Post/properties/id"
        upload-time:
          $ref: "#/components/schemas/DateTime"
        caption:
          type: string
          pattern: '\S.+'
          example: Wow!!!
          minLength: 1
          maxLength: 200
    DateTime:
      title: DateTime
      description: This object reperesents the date and time
      type: object
      readOnly: true
      properties:
        date-time:
          description: The date and time
          type: string
          format: date-time
          minLength: 20
          maxLength: 20
          example: 2017-07-21T17:32:28Z
    Message:
      title: Message
      description: This object represent a generic description of a message
      type: object
      readOnly: true
      properties:
        error:
          description: The message of the error
          type: string
          pattern: '\S.+'
          example: Username is already taken
          minLength: 0
          maxLength: 200
  parameters:
    idPhoto:
      name: idPhoto
      in: path
      schema:
        $ref: "#/components/schemas/Post/properties/id"
      required: true
      description: The id of a photo
    idUser:
      name: idUser
      in: path
      schema: 
        $ref: "#/components/schemas/User/properties/id" 
      required: true
      description: The id of a user
...
  responses:
    OK: # 200
      description: 
        The request succeeded
      content:
        application/json:
          schema:
            description: The message of the error
            type: object
            properties:
              err:
                $ref: "#/components/schemas/Message"
    Created:  # 201
      description: 
        The request succeeded, and a new resource was created as a result
      content:
        application/json:
          schema:
            type: object
            description: The message of the error
            properties:
              err:
                $ref: "#/components/schemas/Message"
...
    BadRequest: # 400
      description: |
        The request sent to the server is invalid (e.g. malformed request syntax, invalid request message framing, ...)
      content:
        application/json:
          schema:
            type: object
            description: The message of the error
            properties:
              err:
                $ref: "#/components/schemas/Message"
    Unauthorized: # 401
      description: 
        The request was not successful because it lacks valid authentication credentials for the requested resource
      content:
        application/json:
          schema:
            type: object
            description: The message of the error
            properties:
              err:
                $ref: "#/components/schemas/Message"
    Forbidden:  # 403
      description: 
        Access denied. You do not have the permission to access this resource or perform this action
      content:
        application/json:
          schema:
            description: The message of the error
            type: object
            properties:
              err:
                $ref: "#/components/schemas/Message"
    NotFound: # 404
      description: |
        The resource you are trying to access doesn't exist
      content:
        application/json:
          schema: 
            type: object
            description: The message of the error
            properties:
              err:
                $ref: "#/components/schemas/Message"
...
    InternalServerError:  #500
      description: The server encounter an unexpected error
      content:
        application/json:
          schema:
            type: object
            description: The message of the error
            properties:
              err:
                $ref: "#/components/schemas/Message"    
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

# Apply the security global, to all the operations
security:
  - bearerAuth: []
...